* master:
  version info: update
  ChangeLog: update
  Bugfix: Do not fail if GetApplicationInfo() return NULL
  Rename HTTP --> WebDAV
  Refactoring
  TStringList::LoadFromFile: use TFileBuffer
  TStringList::LoadFromFile
  FarDialog.cpp: Code cleanup
  WinSCPDialogs.cpp: Code style
  Code style: value --> Value
  Fix warnings
  Fix warning: 'TFarDialogItem::FOnMouseClick' is not initialized in the constructor.
  DWORD --> uintptr_t

Conflicts:
	src/NetBox/FarDialog.cpp
	src/NetBox/FarDialog.h
	src/NetBox/WinSCPDialogs.cpp

+++ src/core/SessionData.h
@@ -183,7 +183,7 @@ private:
   bool FFtpAllowEmptyPassword;
   TLoginType FLoginType;
   int FNumberOfRetries;
-  DWORD FSessionVersion;
+  uintptr_t FSessionVersion;
 
 public:
   void __fastcall SetHostName(UnicodeString value);
